# backend/swagger.json
{"openapi":"3.0.0","paths":{"/user/create":{"post":{"operationId":"UserController_createUser","summary":"Create a new user with images","parameters":[],"requestBody":{"required":true,"description":"User data and images","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateUserRequestDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublicDTO"}}}}},"tags":["User"]}},"/user/{userId}":{"put":{"operationId":"UserController_updateUser","summary":"Update an existing user","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"User data and images","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UpdateUserRequestDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublicDTO"}}}}},"tags":["User"]},"get":{"operationId":"UserController_getUser","summary":"Get a user by ID","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"number"}}],"responses":{"200":{"description":"The user has been successfully retrieved.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found."}},"tags":["User"]}},"/user/location/{userId}":{"put":{"operationId":"UserController_updateLocation","summary":"Update user location","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationUpdateDTO"}}}},"responses":{"200":{"description":"The user location has been successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublicDTO"}}}},"404":{"description":"User not found."}},"tags":["User"]}},"/api/push-token/{userId}":{"post":{"operationId":"NotificationController_storePushToken","summary":"Store user's push token","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StorePushTokenDTO"}}}},"responses":{"201":{"description":"The push token has been successfully stored."},"400":{"description":"Bad Request."},"401":{"description":"Unauthorized."},"500":{"description":"Internal server error."}},"tags":["Push Notifications"]}},"/user-reports/{userId}":{"post":{"operationId":"UserReportController_create","summary":"Create a new user report","parameters":[{"name":"userId","required":true,"in":"path","description":"Reporting User ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserReportDTO"}}}},"responses":{"201":{"description":"The report has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReport"}}}}},"tags":["user-reports"]}},"/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDTO"}}}}},"tags":["Auth"]}},"/encounter/user/{userId}":{"get":{"operationId":"EncounterController_getEncountersByUser","summary":"Get encounters of a user","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EncounterPublicDTO"}}}}},"404":{"description":"Encounter not found."}},"tags":["Encounter"]}},"/registration":{"post":{"operationId":"RegistrationController_registerUserForEmailVerification","summary":"Creates a user with only an email to verify.","parameters":[],"requestBody":{"required":true,"description":"User email.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationForVerificationDTO"}}}},"responses":{"200":{"description":"Email successfully registered. Still needs to be verified.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationForVerificationDTO"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationForVerificationDTO"}}}},"409":{"description":"Email already exists."}},"tags":["Registration"]}},"/registration/verify-email":{"put":{"operationId":"RegistrationController_verifyEmail","summary":"Verify email with verification code.","parameters":[],"requestBody":{"required":true,"description":"User email and verification code.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailDTO"}}}},"responses":{"200":{"description":""}},"tags":["Registration"]}},"/main/uptime":{"get":{"operationId":"AppController_getUptime","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"info":{"title":"Offlinery","description":"API of Offlinery","version":"1","contact":{}},"tags":[{"name":"app","description":""}],"servers":[{"url":"https://offlinery.onrender.com/v1"}],"components":{"schemas":{"BlacklistedRegionDTO":{"type":"object","properties":{"location":{"type":"object","properties":{"type":{"type":"string","enum":["Point"]},"coordinates":{"type":"array","items":{"type":"number"},"minItems":2,"maxItems":2}},"description":"GeoJSON Point representing the location of the blacklisted region (0:lon, 1:lat)","example":{"type":"Point","coordinates":[-74.006,40.7128]}},"radius":{"type":"number","description":"Radius of the blacklisted region in meters","example":1000}},"required":["location","radius"]},"CreateUserDTO":{"type":"object","properties":{"firstName":{"type":"string"},"email":{"type":"string","format":"email"},"clearPassword":{"type":"string"},"wantsEmailUpdates":{"type":"boolean"},"birthDay":{"type":"string","format":"date"},"gender":{"enum":["woman","man"],"type":"string"},"genderDesire":{"enum":["woman","man"],"type":"string"},"verificationStatus":{"enum":["verified","pending","not_needed"],"type":"string"},"approachChoice":{"enum":["approach","be_approached","both"],"type":"string"},"blacklistedRegions":{"description":"Array of blacklisted regions","example":[{"latitude":40.7128,"longitude":-74.006,"radius":1000}],"type":"array","items":{"$ref":"#/components/schemas/BlacklistedRegionDTO"}},"approachFromTime":{"type":"string","format":"date-time"},"approachToTime":{"type":"string","format":"date-time"},"bio":{"type":"string"},"dateMode":{"enum":["ghost","live"],"type":"string"}},"required":["firstName","email","clearPassword","wantsEmailUpdates","birthDay","gender","genderDesire","verificationStatus","approachChoice","approachFromTime","approachToTime","bio","dateMode"]},"CreateUserRequestDTO":{"type":"object","properties":{"user":{"format":"json","allOf":[{"$ref":"#/components/schemas/CreateUserDTO"}]},"images":{"type":"array","items":{"type":"string","format":"binary"},"description":"An array of image files","maxItems":6}},"required":["user","images"]},"UserPublicDTO":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the user"},"firstName":{"type":"string","description":"The first name of the user"},"age":{"type":"number","description":"Age of user"},"imageURIs":{"type":"array","items":{"type":"string"},"description":"An array of image uris","maxItems":6},"bio":{"type":"string","description":"The user's bio"},"trustScore":{"type":"number","description":"The user's trust score"}},"required":["id","firstName","age","imageURIs","bio"]},"UpdateUserDTO":{"type":"object","properties":{"firstName":{"type":"string"},"email":{"type":"string","format":"email"},"wantsEmailUpdates":{"type":"boolean"},"birthDay":{"type":"date","format":"date"},"gender":{"enum":["woman","man"],"type":"string"},"genderDesire":{"enum":["woman","man"],"type":"string"},"verificationStatus":{"enum":["verified","pending","not_needed"],"type":"string"},"approachChoice":{"enum":["approach","be_approached","both"],"type":"string"},"blacklistedRegions":{"description":"Array of blacklisted regions","example":[{"latitude":40.7128,"longitude":-74.006,"radius":1000}],"type":"array","items":{"$ref":"#/components/schemas/BlacklistedRegionDTO"}},"approachFromTime":{"type":"string","format":"time"},"approachToTime":{"type":"string","format":"time"},"bio":{"type":"string"},"dateMode":{"enum":["ghost","live"],"type":"string"},"clearPassword":{"type":"string"}}},"UpdateUserRequestDTO":{"type":"object","properties":{"user":{"format":"json","allOf":[{"$ref":"#/components/schemas/UpdateUserDTO"}]},"images":{"type":"array","items":{"type":"string","format":"binary"},"description":"An array of image files","maxItems":6}}},"Encounter":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string","enum":["not_met","met_not_interested","met_interested"]},"lastDateTimePassedBy":{"format":"date-time","type":"string"},"lastLongitudePassedBy":{"type":"string"},"lastLatitudePassedBy":{"type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"userReports":{"type":"array","items":{"$ref":"#/components/schemas/UserReport"}}},"required":["id","status","lastDateTimePassedBy","lastLongitudePassedBy","lastLatitudePassedBy","users","userReports"]},"UserReport":{"type":"object","properties":{"id":{"type":"number"},"incidentDescription":{"type":"string"},"keepReporterInTheLoop":{"type":"boolean"},"incidentType":{"enum":["Disrespectful","Sexual harassment","Violent behavior","Other"],"type":"string"},"reportedOn":{"format":"date-time","type":"string"},"reportedUser":{"$ref":"#/components/schemas/User"},"reportingUser":{"$ref":"#/components/schemas/User"},"reportedEncounter":{"$ref":"#/components/schemas/Encounter"}},"required":["id","incidentDescription","keepReporterInTheLoop","incidentType","reportedOn","reportedUser","reportingUser","reportedEncounter"]},"User":{"type":"object","properties":{"id":{"type":"string"},"isActive":{"type":"boolean"},"firstName":{"type":"string"},"wantsEmailUpdates":{"type":"boolean"},"email":{"type":"string"},"passwordHash":{"type":"string"},"passwordSalt":{"type":"string"},"birthDay":{"format":"date-time","type":"string"},"gender":{"enum":["woman","man"],"type":"string"},"genderDesire":{"enum":["woman","man"],"type":"string"},"imageURIs":{"type":"array","items":{"type":"string"}},"verificationStatus":{"enum":["verified","pending","not_needed"],"type":"string"},"approachChoice":{"enum":["approach","be_approached","both"],"type":"string"},"blacklistedRegions":{"type":"array","items":{"$ref":"#/components/schemas/BlacklistedRegion"}},"approachFromTime":{"format":"date-time","type":"string"},"approachToTime":{"format":"date-time","type":"string"},"bio":{"type":"string"},"dateMode":{"enum":["ghost","live"],"type":"string"},"pushToken":{"type":"string"},"receivedReports":{"type":"array","items":{"$ref":"#/components/schemas/UserReport"}},"issuedReports":{"type":"array","items":{"$ref":"#/components/schemas/UserReport"}},"encounters":{"type":"array","items":{"$ref":"#/components/schemas/Encounter"}},"trustScore":{"type":"number"},"location":{"type":"object"}},"required":["id","isActive","firstName","wantsEmailUpdates","email","passwordHash","passwordSalt","birthDay","gender","genderDesire","imageURIs","verificationStatus","approachChoice","blacklistedRegions","approachFromTime","approachToTime","bio","dateMode","pushToken","receivedReports","issuedReports","encounters","location"]},"BlacklistedRegion":{"type":"object","properties":{"id":{"type":"number"},"location":{"type":"object"},"radius":{"type":"number"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","location","radius","user"]},"LocationUpdateDTO":{"type":"object","properties":{"latitude":{"type":"number","minimum":-90,"maximum":90,"description":"Latitude of the user's location","example":40.7128},"longitude":{"type":"number","minimum":-180,"maximum":180,"description":"Longitude of the user's location","example":-74.006}},"required":["latitude","longitude"]},"StorePushTokenDTO":{"type":"object","properties":{"pushToken":{"type":"string","description":"The Expo push token for the user's device","example":"ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx]"}},"required":["pushToken"]},"CreateUserReportDTO":{"type":"object","properties":{"incidentDescription":{"type":"string","description":"The description of the incident","example":"The user was being disrespectful in chat"},"keepReporterInTheLoop":{"type":"boolean","description":"Whether to keep the reporter updated on the report status","example":true},"incidentType":{"enum":["Disrespectful","Sexual harassment","Violent behavior","Other"],"type":"string","description":"The type of incident","example":"Disrespectful"},"reportedUserId":{"type":"string","description":"The ID of the user being reported"}},"required":["incidentDescription","keepReporterInTheLoop","incidentType","reportedUserId"]},"SignInDTO":{"type":"object","properties":{"email":{"type":"string"},"clearPassword":{"type":"string"}},"required":["email","clearPassword"]},"UserPrivateDTO":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the user"},"firstName":{"type":"string","description":"The first name of the user"},"age":{"type":"number","description":"Age of user"},"imageURIs":{"type":"array","items":{"type":"string"},"description":"An array of image uris","maxItems":6},"bio":{"type":"string","description":"The user's bio"},"trustScore":{"type":"number","description":"The user's trust score"},"isActive":{"type":"boolean","description":"Indicates if the user account is active"},"email":{"type":"string","description":"The unique email of the user"},"wantsEmailUpdates":{"type":"boolean","description":"Indicates if the user wants to receive email updates"},"blacklistedRegions":{"description":"Locations to not be approached at","type":"array","items":{"$ref":"#/components/schemas/BlacklistedRegionDTO"}},"birthDay":{"type":"string","format":"date"},"approachFromTime":{"type":"string","format":"time","description":"The time from which the user can be approached"},"approachToTime":{"type":"string","format":"time","description":"The time until which the user can be approached"},"dateMode":{"enum":["ghost","live"],"type":"string","description":"The date mode of the user"},"verificationStatus":{"enum":["verified","pending","not_needed"],"type":"string","description":"The verification status of the user"},"approachChoice":{"enum":["approach","be_approached","both"],"type":"string","description":"The approach choice of the user"},"gender":{"enum":["woman","man"],"type":"string","description":"The gender of the user"},"genderDesire":{"enum":["woman","man"],"type":"string","description":"The gender the user is interested in"}},"required":["id","firstName","age","imageURIs","bio","isActive","email","wantsEmailUpdates","blacklistedRegions","birthDay","approachFromTime","approachToTime","dateMode","verificationStatus","approachChoice","gender","genderDesire"]},"SignInResponseDTO":{"type":"object","properties":{"accessToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserPrivateDTO"}},"required":["accessToken","user"]},"EncounterPublicDTO":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the encounter"},"status":{"enum":["not_met","met_not_interested","met_interested"],"type":"string","description":"The status of the encounter"},"lastDateTimePassedBy":{"type":"string","format":"time","description":"The date time they were nearby last time."},"lastLocationPassedBy":{"type":"string","nullable":true,"description":"Last rough location passed by"},"reported":{"type":"boolean","description":"Has this encounter been reported by any of the users"},"users":{"description":"Users that were nearby","type":"array","items":{"$ref":"#/components/schemas/UserPublicDTO"}}},"required":["id","status","lastDateTimePassedBy","reported","users"]},"RegistrationForVerificationDTO":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"}},"required":["email"]},"VerifyEmailDTO":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"verificationCode":{"type":"string","minLength":6,"maxLength":6,"example":"123456"}},"required":["email","verificationCode"]},"NotificationNavigateUserDTO":{"type":"object","properties":{"screen":{"type":"string","enum":["Main_NavigateToApproach"]},"navigateToPerson":{"$ref":"#/components/schemas/UserPublicDTO"}},"required":["screen","navigateToPerson"]}}}}